+++ src/core/GrantList.ts
@@ -1,0 +1,5 @@
+// own modules
+import { IGrantListItem } from './IGrantListItem';
+
+/** Grant list to pass AccessControl's constructor */
+export type GrantList = IGrantListItem[];